Split motion request builder into its own file
