+++ robowflex_library/include/robowflex_library/robowflex.h
@@ -29,6 +29,7 @@ namespace robowflex
 #include <robowflex_library/scene.h>
 #include <robowflex_library/robot.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/benchmarking.h>
 #include <robowflex_library/path.h>
