+++ robowflex_library/scripts/r2_test.cpp
@@ -3,6 +3,7 @@
 #include <robowflex_library/robot.h>
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/io/visualization.h>
 #include <robowflex_library/detail/r2.h>
