+++ robowflex_library/scripts/ur5_test.cpp
@@ -5,6 +5,7 @@
 #include <robowflex_library/robot.h>
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/detail/ur5.h>
 
 using namespace robowflex;
