+++ robowflex_library/scripts/ur5_visualization.cpp
@@ -6,6 +6,7 @@
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
 #include <robowflex_library/io/visualization.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/detail/ur5.h>
 
 using namespace robowflex;
