+++ robowflex_library/scripts/wam7_benchmark.cpp
@@ -3,6 +3,7 @@
 #include <robowflex_library/robot.h>
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/benchmarking.h>
 
 using namespace robowflex;
