+++ robowflex_library/scripts/wam7_test.cpp
@@ -2,6 +2,7 @@
 #include <robowflex_library/geometry.h>
 #include <robowflex_library/robot.h>
 #include <robowflex_library/scene.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/planning.h>
 
 using namespace robowflex;
