+++ robowflex_library/src/benchmarking.cpp
@@ -10,6 +10,7 @@
 #include <robowflex_library/path.h>
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/benchmarking.h>
 
 using namespace robowflex;
