+++ robowflex_ompl/scripts/ur5_ompl_interface.cpp
@@ -5,6 +5,7 @@
 #include <robowflex_library/robot.h>
 #include <robowflex_library/scene.h>
 #include <robowflex_library/planning.h>
+#include <robowflex_library/builder.h>
 #include <robowflex_library/detail/ur5.h>
 
 #include <robowflex_ompl/ompl_interface.h>
